+++ src/shared/chat/message-body.js
@@ -11,12 +11,12 @@ export default class MessageBody extends CustomElement {
 
     static get properties () {
         return {
-            embed_audio: { type: Boolean },
-            embed_videos: { type: Boolean },
-            hide_url_previews: { type: Boolean },
+            // We make this a string instead of a boolean, since we want to
+            // distinguish between true, false and undefined states
+            hide_url_previews: { type: String },
             is_me_message: { type: Boolean },
             model: { type: Object },
-            show_images: { type: Boolean },
+            render_media: { type: Boolean },
             text: { type: String },
         }
     }
@@ -33,19 +33,25 @@ export default class MessageBody extends CustomElement {
     render () {
         const callback = () => this.model.collection?.trigger('rendered', this.model);
         const offset = 0;
-        const mentions = this.model.get('references');
-
         const options = {
-            'embed_audio': !this.hide_url_previews && this.embed_audio,
-            'embed_videos': !this.hide_url_previews && this.embed_videos,
+            'media_urls': this.model.get('media_urls'),
+            'mentions': this.model.get('references'),
             'nick': this.model.collection.chatbox.get('nick'),
             'onImgClick': (ev) => this.onImgClick(ev),
             'onImgLoad': () => this.onImgLoad(),
             'render_styling': !this.model.get('is_unstyled') && api.settings.get('allow_message_styling'),
-            'show_images': !this.hide_url_previews && this.show_images,
-            'show_me_message': true
+            'show_me_message': true,
         }
-        return renderRichText(this.text, offset, mentions, options, callback);
+        if (this.hide_url_previews === "false") {
+            options.embed_audio = true;
+            options.embed_videos = true;
+            options.show_images = true;
+        } else if (this.hide_url_previews === "true") {
+            options.embed_audio = false;
+            options.embed_videos = false;
+            options.show_images = false;
+        }
+        return renderRichText(this.text, offset, options, callback);
     }
 }